+++ tests/tests/tests_commands.py
@@ -7,20 +7,3 @@ class TestDeleteMigrations(TestCase):
 
     def test_call_command(self):
        call_command('deletemigrations', 'tests')
-
-
-class TestStartAppPlus(TestCase):
-
-    def test_call_command(self):
-        call_command('startapp_plus', 'app_for_test')
-
-    @classmethod
-    def tearDownClass(self):
-        import os
-        from shutil import rmtree
-        from django.conf import settings
-
-        for f in ['app_for_test']:
-            fp = os.path.join(settings.BASE_DIR, f)
-            if os.path.exists(fp):
-                rmtree(fp)